requirements.txt
# WutheringWaves-Echo-OCR 依赖包

# 核心依赖（版本锁定确保兼容性）
numpy<2.0.0
onnxruntime>=1.19.0

# GUI 框架
PySide6>=6.6.0

# OCR 引擎
rapidocr-onnxruntime>=1.3.0

# 图像处理（锁定版本避免兼容性问题）
opencv-python==4.8.0.76
Pillow>=10.0.0

# 数据处理
PyYAML>=6.0

# 开发工具
pytest>=7.4.0
pytest-qt>=4.2.0
